medinovai-deployment/services/healthllm/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
httpx==0.25.2
python-dotenv==1.0.0
psycopg2-binary==2.9.9
redis==5.0.1

